fix suba for our headers

+++ src/cpu/m68000/m68k_in.c
@@ -9482,7 +9482,7 @@ M68KMAKE_OP(suba, 16, ., a)
 M68KMAKE_OP(suba, 16, ., .)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(M68KMAKE_GET_OPER_AY_16);
+	signed short src = MAKE_INT_16(M68KMAKE_GET_OPER_AY_16);
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
