+++ src/cpu/m68000/m68kopnz.c
@@ -6232,7 +6232,7 @@ void m68k_op_suba_16_a(void)
 void m68k_op_suba_16_ai(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_AY_AI_16());
+	signed short src = MAKE_INT_16(OPER_AY_AI_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6240,7 +6240,7 @@ void m68k_op_suba_16_ai(void)
 void m68k_op_suba_16_pi(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_AY_PI_16());
+	signed short src = MAKE_INT_16(OPER_AY_PI_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6248,7 +6248,7 @@ void m68k_op_suba_16_pi(void)
 void m68k_op_suba_16_pd(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_AY_PD_16());
+	signed short src = MAKE_INT_16(OPER_AY_PD_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6256,7 +6256,7 @@ void m68k_op_suba_16_pd(void)
 void m68k_op_suba_16_di(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_AY_DI_16());
+	signed short src = MAKE_INT_16(OPER_AY_DI_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6264,7 +6264,7 @@ void m68k_op_suba_16_di(void)
 void m68k_op_suba_16_ix(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_AY_IX_16());
+	signed short src = MAKE_INT_16(OPER_AY_IX_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6272,7 +6272,7 @@ void m68k_op_suba_16_ix(void)
 void m68k_op_suba_16_aw(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_AW_16());
+	signed short src = MAKE_INT_16(OPER_AW_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6280,7 +6280,7 @@ void m68k_op_suba_16_aw(void)
 void m68k_op_suba_16_al(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_AL_16());
+	signed short src = MAKE_INT_16(OPER_AL_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6288,7 +6288,7 @@ void m68k_op_suba_16_al(void)
 void m68k_op_suba_16_pcdi(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_PCDI_16());
+	signed short src = MAKE_INT_16(OPER_PCDI_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6296,7 +6296,7 @@ void m68k_op_suba_16_pcdi(void)
 void m68k_op_suba_16_pcix(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_PCIX_16());
+	signed short src = MAKE_INT_16(OPER_PCIX_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
 
@@ -6304,7 +6304,7 @@ void m68k_op_suba_16_pcix(void)
 void m68k_op_suba_16_i(void)
 {
 	uint* r_dst = &AX;
-	uint src = MAKE_INT_16(OPER_I_16());
+	signed short src = MAKE_INT_16(OPER_I_16());
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
